+++ modules/juce_audio_devices/native/juce_win32_AudioCDReader.cpp
@@ -1069,7 +1069,7 @@ bool AudioCDReader::readSamples (int** destSamples, int numDestChannels, int sta
         }
         else
         {
-            const int framesInBuffer = buffer.getSize() / bytesPerFrame;
+            const int framesInBuffer = (int) (buffer.getSize() / bytesPerFrame);
             const int frameNeeded = (int) (startSampleInFile / samplesPerFrame);
 
             if (firstFrameInBuffer + framesInBuffer != frameNeeded)
