+++ modules/juce_audio_plugin_client/RTAS/juce_RTAS_Wrapper.cpp
@@ -848,7 +848,7 @@ private:
             // Pro-tools expects all your parameters to have valid names!
             jassert (juceFilter->getParameterName (index).isNotEmpty());
 
-            juceFilter->getParameterName (index).copyToUTF8 (name, maxLength);
+            juceFilter->getParameterName (index).copyToUTF8 (name, (size_t) maxLength);
         }
 
         long GetPriority() const        { return kFicCooperativeTaskPriority; }
@@ -863,7 +863,7 @@ private:
 
         void GetValueString (char* valueString, int maxLength, long value) const
         {
-            juceFilter->getParameterText (index).copyToUTF8 (valueString, maxLength);
+            juceFilter->getParameterText (index).copyToUTF8 (valueString, (size_t) maxLength);
         }
 
         Cmn_Bool IsAutomatable() const
